Add screens for baseline profile generator
