+++ benchmarks/src/main/kotlin/com/merxury/blocker/baselineprofile/RuleBaselineProfile.kt
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2023 Blocker
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.merxury.blocker.baselineprofile
+
+import androidx.benchmark.macro.junit4.BaselineProfileRule
+import com.merxury.blocker.PACKAGE_NAME
+import com.merxury.blocker.rules.goToRulesScreen
+import com.merxury.blocker.rules.rulesScrollListDownUp
+import com.merxury.blocker.startActivityAndAllowNotifications
+import org.junit.Rule
+import org.junit.Test
+
+/**
+ * Baseline Profile of the "Rules" screen
+ */
+class RuleBaselineProfile {
+    @get:Rule val baselineProfileRule = BaselineProfileRule()
+
+    @Test
+    fun generate() =
+        baselineProfileRule.collect(PACKAGE_NAME) {
+            startActivityAndAllowNotifications()
+
+            // Navigate to the rules screen
+            goToRulesScreen()
+            rulesScrollListDownUp()
+        }
+}
